+++ src/drink.py
@@ -1,0 +1,7 @@
+class Drink:
+    """Represents a drink item in the vending machine."""
+
+    def __init__(self, name: str, price: int, count: int) -> None:
+        self.name = name
+        self.price = price
+        self.count = count
